Check at interval and show notification
Add global state in main.zig so that the allocator, menu and client is
accessible in the callback functions.
Do a periodic check for new notifications and show a popup when
notifications happen.
lookup the users preferred browser in the registry for opening the repo
page or notifications on naturbasen.